+++ src/main/java/club/inq/team1/constant/Gender.java
@@ -1,4 +1,4 @@
-package club.inq.team1.config;
+package club.inq.team1.constant;
 
 public enum Gender {
     FEMALE,MALE
