+++ src/main/java/club/inq/team1/dto/request/user/RequestUserCreateDTO.java
@@ -1,6 +1,6 @@
 package club.inq.team1.dto.request.user;
 
-import club.inq.team1.config.Gender;
+import club.inq.team1.constant.Gender;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
