+++ src/main/java/club/inq/team1/repository/post/PostRepository.java
@@ -1,12 +1,10 @@
-package club.inq.team1.repository;
+package club.inq.team1.repository.post;
 
 import club.inq.team1.entity.Post;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 public interface PostRepository extends JpaRepository<Post, Long> {
