+++ src/main/java/club/inq/team1/service/impl/post/CommentServiceImpl.java
@@ -7,8 +7,8 @@ import club.inq.team1.entity.Comment;
 import club.inq.team1.entity.CommentLike;
 import club.inq.team1.entity.Post;
 import club.inq.team1.entity.User;
-import club.inq.team1.repository.CommentRepository;
-import club.inq.team1.repository.PostRepository;
+import club.inq.team1.repository.post.CommentRepository;
+import club.inq.team1.repository.post.PostRepository;
 import club.inq.team1.repository.post.CommentLikeRepository;
 import club.inq.team1.service.post.CommentService;
 import club.inq.team1.service.post.ReplyService;
