+++ src/main/java/club/inq/team1/service/impl/user/FollowServiceImpl.java
@@ -4,9 +4,9 @@ import club.inq.team1.dto.response.user.ResponseUserPrivateInfoDTO;
 import club.inq.team1.entity.Follow;
 import club.inq.team1.entity.Mail;
 import club.inq.team1.entity.User;
-import club.inq.team1.repository.FollowRepository;
-import club.inq.team1.repository.MailRepository;
-import club.inq.team1.repository.UserRepository;
+import club.inq.team1.repository.user.FollowRepository;
+import club.inq.team1.repository.user.MailRepository;
+import club.inq.team1.repository.user.UserRepository;
 import club.inq.team1.service.user.FollowService;
 import club.inq.team1.service.user.UserService;
 import club.inq.team1.util.CurrentUser;
