+++ src/main/java/club/inq/team1/service/impl/user/UserServiceImpl.java
@@ -8,8 +8,8 @@ import club.inq.team1.dto.response.user.ResponseUserPrivateInfoDTO;
 import club.inq.team1.dto.response.user.ResponseUserPublicInfoDTO;
 import club.inq.team1.entity.User;
 import club.inq.team1.entity.UserInfo;
-import club.inq.team1.repository.UserInfoRepository;
-import club.inq.team1.repository.UserRepository;
+import club.inq.team1.repository.user.UserInfoRepository;
+import club.inq.team1.repository.user.UserRepository;
 import club.inq.team1.service.user.UserService;
 import club.inq.team1.util.CurrentUser;
 import jakarta.transaction.Transactional;
